create funtoken_from_coin_test.go
